[MOVEONLY] Move Get{Sequence,Output,Prevout}Hash  to primitives/transaction.h
